ESE Block Cache:  Performance:  optimize sync read used to determine caching state on file open
EBC currently performs an uncached read at offset 0 on every file open to determine if a given file is cached.

This change introduces a way to mark a file as having ever been considered for caching.  This mark will allow EBC to skip this read for any file that has never been considered for caching by EBC.

The mark is implemented via an NTFS alternate data stream.  This ADS is stamped on any file that is opened by EBC when its configuration indicates that caching is desired (ICachedFileConfiguration::FCachingEnabled).  The existence of the ADS then determines if we should continue on to attempt to find the cached file header.  We chose to use ADS for this purpose because it is supported by NTFS and by ReFS and its representation as a path should allow it to be emulated if necessary by other file systems.  It is also supported by Windows SMB.

New test cases were added for COSFileSystem/COSFile to ensure that they support the required ADS semantics.  Surprisingly, no code changes were required here.

This is a format change but it is implemented such that the existing cached files in production will be stamped before open (because their ICachedFileConfiguration::FCachingEnabled is true) so the upgrade will occur naturally.

This change will prevent EBC from causing extra useless HDD IO in production for HDD based files.  This will be most impactful for log files that are not cached on flash drives, mostly legacy lag copies.

This change also defers the open of the cache configured to back a cached file until it is actually used.  Previously we were creating/opening the cache even if we didn't actually use it.

[Substrate:0ac4e4075c25c71b86df602ea8ef75d7b7a76811]