Adding an auto-inc column with a default value fails with JET_errInvalidParameter.
We've recently added the ability to set a value for an auto-increment column explicitly, via JET_prepInsertMustSetAutoIncrement. Along with that feature, we've added checks to make sure invalid values aren't set, which then causes JetSetColumn() to return either JET_errInvalidParameter or JET_errSetAutoIncrementTooHigh, depending on the case.

These extra checks broke one of our clients, because they create an auto-inc column with a default value. When the value is then internally obtained and attempted to be set on the column, it might hit a failure condition and fail API calls. Defining an auto-inc column with a default value does not make sense, because auto-inc columns are always internally set with the next auto-incremented value and the user-set default value always ends up being clobbered anyways. Unfortunately, this invalid usage hasn't been blocked by our API in the past, so we now need to handle it.

The fix:
1. Only check values being set on auto-inc columns if we are under a session that has been prepared with JET_prepInsertMustSetAutoIncrement.
2. Block using JET_prepInsertMustSetAutoIncrement with an auto-inc column that has a default value. This is necessary to avoid having a client bypass internally setting the auto-inc column with a consistent value and proceed to complete the update without having explicitly set a value, since we can't detect that the column hasn't been set by the client (rather than set internally, which it did, by the default value) .
3. Block using JET_TblInfoRetrieveAndReserveAutoIncrement with an auto-inc column that has a default value. This is not strictly necessary, but because this is the expected way to get a consistent auto-inc value to be set with explicitly with JET_prepInsertMustSetAutoIncrement, we'll block it for safety and API consistency.

[Substrate:6bcb55291e9d6fd2b936c5dde162eb9b5194bbac]