ESE Block Cache:  cluster management corruption
Stress testing has very rarely hit a failure where the same cluster ended up assigned to more than one slot.  I recently successfully captured one with enough state to discover the problem.

EBC uses "transactions" to perform some operations.  A transaction is implemented as an update of more than one chunk of a slab in the same journal entry.  These trx are primarily used to evict/invalidate the contents of a slot.  When this happens, we must swap out the current cluster holding data for that slot with a new one from a journal slab.  This facilitates our ability to overwrite new data in the cache as a virtual part of the journal w/o damaging previous cache state in case the update is unsuccessful and dropped on recovery by journal truncation.

The stress test is configured to cache all ESE files.  In one case, we hit a file sharing error trying to open a file backing data we wanted to evict.  This error happened after we had already evicted some data in the slab and swapped one cluster.  The error handling caused us to abandon the update on the journal cluster slab but not the partial update on the normal slab.  This resulted in that cluster being double allocated and the old cluster being lost.  A subsequent update then caused data corruption because two slots were pointing to the same cluster.

This failure revealed many things, all of which were addressed:

- an ErrUnexpectedDataReadFailure doesn't emit an event or a failure item enabling page patching
- the cache must not evict data for a cached file unless it fails to access it because it is conclusively deleted
- the cache should not fail when cleaning a slab due to a transient error when opening a cached file
- any failure during an update must not allow partial state to be committed
- any given journaled update must not allow partial cluster swaps

Test cases were added to explicitly exercise the failure cases we encountered in the stress test.  They are confirmed to fail in the expected way on the previous code.

Supporting changes:

- fixed a crash during CRequest completion when tracing was enabled that resulted in a leak
- a variant of OSFileIIOReportError that can be used by EBC was exposed internally to the OS Layer
- ICachedBlockSlab::ErrUpdateSlot was tweaked to guarantee the success of an undo of an update

[Substrate:298309ee75d6d5fc5e07dc421e3d04fbf707f751]