Provide mechanism to read various symmetric page usages
Added ability to read avail and owned for the LV tree inside another table.
Added ability to read avail and owned for indices.  This reads exclusive values (i.e. owned for primary index, which is the same FCB as the table, returns only pages owned by the table that are not owned by 2ndary indices, etc).

[Substrate:a815f95f43574a9e086def5ecb292573febf8777]