ESE Block Cache:  In Place Destage
This change implements the ability to write back all data to a cached file and detach that file from the cache.  This process can be done directly via eseutil or indirectly via setting override.  Note that this is different from the existing out of place mechanism we currently have which is essentially a logical file copy.  Also note that the primary mode that will be used in production is driven by setting changes.

In place destage can occur when a cached file is opened.  At that time, we will decide if the criteria for caching that file have changed enough that the file should be detached from the current cache.  This can be simple things like caching has been disabled or it can be more complex like the intended cache implementation or size has changed.

If the detach operation is unsuccessful then the open of the file will succeed and it will continue to use the cache.  Note that detach will fail if it gets a read verification error from the cache.  If this happens in prod the hope is that a page patch will fix the database so that we can try again later.

The detach operation can take a significant amount of time for large caches.  As such, we must control when a detach may occur even if the settings dictate it should occur.  Detach must be explicitly enabled via a system parameter or via configuration so that a non-EBC aware application doesn't inadvertently try to detach a file.  In prod, we will only allow a detach of a cached EDB in store.worker during (usually) passive mount.  We specifically wish to avoid a detach while HA is inspecting the health of a database copy.

The destage process is journaled and is restartable.  The detach process requires a header overwrite and can fail.

Backwards compatible format changes:

- The cached file header now contains a persistent indicator for the size of the pinned header.  This is required to allow reliable destage of the files.  If the field is zero then we interpret that to be 64kb to match the setting for the existing cached files in production.
- The hashed lruk cache header now contains a persistent indicator of its configured max size (vs its actual size) to allow us to determine reliably if the configured cache size has changed.  If the field is zero then it returns the actual size.  We check this with a tolerance of 0.1% which should allow the existing caches in production to think they are sized as configured.

Incomplete work:

- There is ongoing discussion as to how eseutil will provide detach.  This change implements it as a /detach flag for the copy mode (eseutil /y) as an example
- We want to allow "rolling upgrades" of one type or size of cache to another via detach.  To accomplish this, we need to detect when a cache is empty and may be deleted and make it delete itself on close.  This isn't yet implemented.
- This does not implement automatic detach of files because they are entirely kicked out of the cache.

[Substrate:7ddf7cf3df0158874c7563aff0c4b6fa567856b6]