ESE Block Cache:  fix CioNonFlushed() discrepancies caused by internal EBC IO
There is code in ESE that asserts that it has properly flushed a file before it is closed etc.  This code presumes it is safe to check this because no other IO is happening.  This isn't true if EBC is doing write back to the file.

This change implements a separate count for unflushed IO in TFileFilter for engine IO only to allow ESE to get data that is accurate from its point of view.

[Substrate:8280b26508ad5eb9b0f2a1e2320674130019bcd8]