ESE Block Cache:  Presence Filter
This change greatly increases the performance of cache misses by providing a very cheap means to determine that a particular access is not cached.  This is implemented via a probabilistic data structure that can know when an item is definitely not cached or if it is possibly cached (i.e. false negatives are not possible but false positives are possible).

We have chosen to use a Cuckoo Filter to implement the presence filter.  See https://github.com/efficient/cuckoofilter for the paper and its sample implementation.  A Cuckoo Filter uses about as much space as a Bloom Filter, supports duplicate entries and deletes, and is far simpler to implement.  We currently use 8 bits per 4kb cached block in the cache which should yield a 3% false positive rate.  If this is insufficient then we can improve this in the future.  We did not use all the available tricks to keep complexity down.  The code is already factored to allow improvements here though.

Also in the spirit of initial simplicitly, we did not harden this data structure against corruption.  There are checks that will Enforce if corruption is obvious.  We will rely on the higher level data integrity protection in ESE as the guard against corruption.  This can also be easily addressed if needed.

We defer load the contents of the presence filter by slab as they are first referenced.  This is tracked by another simple bit map (1 bit per slab, currently 104 * 4kb of cache).  If the slab's info isn't yet loaded into the PF then we will simply claim the cluster is cached as false positives are ok.  The contents of the PF are updated in sync with journaled updates to the slabs.

The PF can also be used in the future to enable cheap probing to allow write combining across slabs for async write back.  This would allow us to build the same write combining we already have in BF for EBC.

Supporting changes:

-  NextPowerOf2 was improved to use a better algorithm and to support 64 bit numbers
-  ICachedBlockSlab::ErrVisitSlots was improved to expose a CCachedBlockSlotState for the accepted state
-  The slab now indicates if an accepted slot is superceded
-  The slot state now indicates if an update is the first one (UpdateNumber == 1)
-  The slab no longer returns stale data (e.g. updno) for invalid entries.  It returns a default slot
-  The mixed tabulation hash was factored out of the slab manager to allow reuse

[Substrate:e244bb64324ea9589679ac68ac35205acad442d6]