ESE Block Cache:  fix throttling assert in debug stress
The throttling code assumed that if a wait was required then we must need to wait for a concurrent IO count and that we must have failed to combine with an existing one.  There is a race condition where we are setup to wait for a concurrent IO count and then we combined with an existing IO, causing us to cancel our wait but another thread has already granted the concurrent IO count.  The other thread can end up losing its quantum before actually calling CWaiter::Complete.  This leaves us in a state where !CWaiter::FComplete but !fThrottledAcquired and fCombined.  This causes the assert to fire but when inspected, CWaiter::FComplete.

The code should only enter the wait block (where we call ErrIOIssue etc) if fThrottleAcquired && !fCombined && !CWaiter::FComplete.  However, we still need to call waiter.Wait() to handle the race where the waiter hasn't been released just yet.

We also simplified the "self release" case in GetIOWaitersToComplete to directly release the waiter which is on our own thread.  We can do this in the critsec because there is no one waiting so priority inversion penalties for calling CWaiter::Complete are not possible.

[Substrate:08d60a655c8c84d719430b762db5d236567d0a3c]