Add EseIO telemetry for BF writing a new page
This change adds a new IOREASONFLAG to indicate when BF is writing a page that was initially cached as a new page.  This telemetry should point out pages that are written that don't usually need to be read during redo.  The goal is further general understanding but also to inform a particular IO optimization for the ESE Block Cache where we try to not cache the initial write of a new page to avoid cache pollution due to bulk data loads.

We had already used all 8 bits of IOREASONFLAG so we retired iorfShadow.  The value of this flag in the telemetry was limited.

[Substrate:43908d6127ee6b3732d17c9ce4894ce5117989f3]