ESE Block Cache:  AsyncSlabWriteBackWorker performance
There were some bugs in this worker that effectively caused us to always think the replay pointer was too deep, causing us to very aggressively flush our metadata.  This was causing this worker to consume 3.5% CPU in prod which is way higher than expected.  A lot of this time was spent force flushing the journal (append a journal entry and flush) and the file system as well.  We also spent a bunch of time repeatedly flushing our write counts (used to detect lost writes on slab metadata) which can be sizable for a large cache (15GB cache -> 140KB write counts)).

The fixes include:

- change CJournalWrapper::ErrFlush to ignore any request that doesn't advance the durable for write back or the durable pointer (i.e. a nop flush)
- fix the replay pointer target computation
- fix the code that flushes the journal to allow slabs to be written back to be much more conservative and only try to flush to guarantee progress instead of flushing to guarantee all candidates are written back

This great decrease in flushes caused us to discover a problem with our recovery scheme.  Previously we would track invalid Cluster References in the Journal (to detect the tail or corruption) by watching only updates that had Cluster References and ensuring that at least the last one is valid.  This doesn't work if a Cluster is used, the Journal entry is emitted but is not made durable, and is then lost on a remount because we no longer know if that Cluster was reused.  The fix is to remember invalid Cluster References and then forget them if the slot holding that Cluster is evicted or invalidated.  This works because a cluster cannot be reused until its release from its previous use is durable.

This great decrease in flushes also exposed some issues caused by non-durable updates on cache release.  ESE doesn't always flush the file before it closes it.  We changed the cache repository to call a new ICache::ErrPrepareToDismount when a cache is released.  The HashedLRUKCache will now perform a journal flush at this time.  This could be extended to a full state flush in the future if needed.

We also had to fix a bug where we tried to write back slabs during mount, hit a journal full condition, and failed to handle the journal full condition because we couldn't fully truncate the journal because we were still replaying journal entries.

Side effects may include:

- lower write IOPS on the flash drives
- longer recovery times which could lead to longer times to open a file if the cache isn't already open
- more journal full recovery operations (for which we have telemetry)
- more interesting journal entries to dump when the cache is offline
- slabs will be -1119 (lost write) more often in the offline dump, complicating analysis

[Substrate:84a192efe866fc9f36b633e6f5b0673d6ad38ce9]