ESE Block Cache:  fix nop flush detection
EBC tries to avoid flushing the journal if it won't actually improve our durability situation.  This logic has a hole where a previous flush (which emits a journal entry) makes us think that there is unflushed data.

The change is to only consider previous appends that are not flushes for this logic.

[Substrate:9789c1d779c3459190801657c0bcf3125c9c18d9]