ESE Block Cache:  Cache dump improvements
This fixes some problems with the cache text dump via eseutil /mb found while debugging an issue:

- the spacing around the write counts dump is messed up
- the dump stops at a corrupt slab rather than dumping the state for easy diagnosis

This change also includes improvements for Factory.LoadCachedBlockSlab etc so that it can be used to load the binary image of a slab to facilitate debugging.  Specifically, it allows the return of a slab that is in a failed state along with the error just like in ICachedBlockSlabManager.GetSlab.

[Substrate:9185217adb58acc2c2aa15a915bc3543ae2e7e02]