ESE Block Cache:  slot reuse strategy can cause stale updno which can in turn cause lost writes
In prod, we have had two serious cache failures.  One was a case where we lost two consecutive writes to one block.  The second was a case where we failed to update a database header because of a duplicate updno failure.  We didn't have much to debug for the first case but the second case revealed an interesting problem.  Our management of the slots in a slab has an issue where we could leave slots with very old state unused for a very long time.  This could result in a set of updnos for a given cached block that were so far apart that the wrap around aware comparison of the 16 bit updno fails and declares the oldest slots as the youngest.  This causes our state management to get confused.  The cache can then start returning old data as if it were current.

To resolve this issue, we have changed our slot management to preferentially use the least recently used slot when we need to allocate a new slot.  We will use the last touch time (tono0) for this purpose.  This scheme should guarantee that the updnos for a given cached block never span more than half the precision of the updno.

To support this, we now retain tonos on slots that are no longer valid.

[Substrate:c0aa7d2cacc5ee831e9f4fb970f04c2e76c79f45]