ESE Block Cache:  pooled allocations
EBC has a lot of code paths that allocate blocks of VM to perform IO.  This is expensive.

This change includes a new pool implementation that can be used to retain a small amount of recently allocated VM blocks for reuse.  It will retire blocks that haven't been used for some min interval of time.  The pool is constructed such that we conspire to share the pool amongst as many compatible types as possible.  For example, all 4kb types end up sharing the same pool.  The one exception is the slab pool (which is actually the hottest) which must have its own pool because its size is determined via variant configuration.

The pool also has an explicit opt out for zeroing VM on allocation if it is reused.  We can use this for certain hot paths (again, the slab pool) if we know we are just going to read into the buffer anyway.

The pool has also been used to (yet again) fix the scheme we use in CIOComplete to avoid the hazard where BF can erroneously access the IO context of an IO after it has completed.  We now use this pool to guarantee that any CIOComplete we give out will live for at least 10s after it is released.  This will dodge the race in BF.  It should be much cheaper than the v2 solution of registering all CIOComplete instances in a hash table which is surprisingly expensive.

[Substrate:04ff686102ec1b4750b7167ef36d5ce5846287df]