ESE Block Cache:  disable leak detection for ExchangeDatabaseManager due to .NET Core issue
We have been fighting with the native resource cleanup in EDM for a while.  I finally got it to work for .NET but now it is failing again in .NET Core.  Previously, we would do the cleanup on AppDomain.CurrentDomain.DomainUnload.  In .NET Core, AppDomain.CurrentDomain.ProcessExit is triggered and we do call Cleanup() but something goes wrong and the process exits and unloads ESE.DLL without releasing the resources, tripping our memory leak detection which crashes the process.

We can reproduce this at will with the MailboxFlags test.  However, whenever we try to debug it, the debugger crashes.  We tried to use TTD as well but it takes so long that the test times out.  So the real failure is a mystery.

To unblock the .NET Core port, we will disable the memory leak detection for just EDM and only when EBC is enabled.  This will leave leak detection on for all the core ESE tests.

[Substrate:3d03c01c539144aa2f6639051c5238a58f4e6022]