O365Core # 2837122 - Make page patching of lost flush crash consistent.
Online page patching flips the state of the flush map to pgftUnknown prior to issuing the sync write to patch the pages and stamps the actual flush types once the write completes. This leads to a problem when we are patching lost flushes and the patching does not complete successfully, because the map state is left at pgftUnknown, which thwarts our ability to detect the lost flush again. We then open ourselves to a bad update in the future which, in turn, may lead to DbTimeTooNew when replaying the log stream.

The fix is to leave the flush map untouched when preparing the write and, instead, stamp that flush type into the page to be patched.

[Substrate:54e79fe447cc7484fedc0f6b0ce94d6adc7141c4]